Updated FPS Movement Function and signs var
Signs seemed right to me. just the fpsMovement seemed to have some errors in it based on what you tell tankControls to do in its conditionals.
May be wrong/blind though, which is why I'm making this a pull request before integrating it.
I made comments on each thing i changed in the program for you to track/approve.